Moved NativeArrayConverter to NativeArray namespace
